[Refactor] Refactor and Remove css style
