+++ src/main/resources/templates/login/index.html
@@ -68,7 +68,7 @@
 
 <script src="login/vendor/jquery/jquery-3.2.1.min.js"></script>
 <script src="login/vendor/animsition/js/animsition.min.js"></script>
-<script src="login/vendor/bootstrap/js/popper.js"></script>
+<script src="login/vendor/bootstrap/js/popper.min.js"></script>
 <script src="login/vendor/bootstrap/js/bootstrap.min.js"></script>
 <script src="login/js/main.js"></script>
